requirements.txt
streamlit==1.50.0
google-generativeai==0.8.5
firebase-admin==7.1.0
Pillow==11.3.0
pandas==2.3.3
plotly==6.3.1
twilio==9.8.4
statsmodels==0.14.5
